Finshed work at Chat agent
Agents should be created in game execution when an specific "login"
event is received.
The rest of the events will be redirected to the agent that handles the
sender IP. In the other way, GameManagers that send events to the chat
agent will be redirected to their clients (using ips).